feat!: merge subscribe and update, wip
