+++ test/storage.broadcast.test.js
@@ -2,8 +2,7 @@
 
 import ninos from "ninos";
 import ava from "ava";
-import { Storage } from "../src/storage";
-import { updateData, updateAndSend } from "../src/update";
+import { Storage, updateData, updateAndSend } from "../src";
 import { args, unhandledMessageError } from "./util";
 
 const test = ninos(ava).serial;
